split interpolation strategy into separate classes with a codec
